Update Selector.java
Added a check when loading the database to make sure the data types match what is expected.

If the data types don't match at any point, it will automatically erase the database file and make a new one.

This should fix the issue that @b16r05 was having with the program (hopefully!)