Update Searcher.java
Removed commented out MINIMUM_NUMBER_OF_UPVOTES as it is now user defined

Added a json check for if a post is an image. All non images are skipped! This fixes an issue where it would download a webpage or video as a png and set it as the wallpaper (black)

Removed video check code as it is handled much better the new way

Removed redundant gallery variable

Added comments to make it more clear what is happening now inside the imageSize() method